# app/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SecureBank API",
    "description": "Comprehensive banking management system with Cedar-based authorization, role-based access control, account management, transactions, loans, and user administration.",
    "version": "1.0.1",
    "contact": {
      "name": "SecureBank API Support",
      "url": "http://localhost:5000"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string", "enum": ["Account Holder", "Bank Teller", "Bank Manager"] }
                },
                "required": ["email", "role"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "operationId": "logout",
        "security": [{"sessionAuth": []}],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "operationId": "getCurrentUser",
        "security": [{"sessionAuth": []}],
        "responses": {
          "200": {
            "description": "Current user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "operationId": "getUsers",
        "security": [{"sessionAuth": []}],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "operationId": "createUser",
        "security": [{"sessionAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId",
        "description": "Session-based authentication via cookies"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "John Smith" },
          "email": { "type": "string", "format": "email", "example": "john.smith@email.com" },
          "phone": { "type": "string", "example": "555-0101" },
          "role": { "type": "string", "enum": ["Account Holder", "Bank Teller", "Bank Manager"] },
          "status": { "type": "string", "enum": ["active", "inactive", "suspended"] },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "email", "role", "status", "createdAt"]
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Emma Davis" },
          "email": { "type": "string", "format": "email", "example": "emma.davis@email.com" },
          "phone": { "type": "string", "example": "555-0104" },
          "role": { "type": "string", "enum": ["Account Holder", "Bank Teller", "Bank Manager"] },
          "status": { "type": "string", "enum": ["active", "inactive", "suspended"], "default": "active" }
        },
        "required": ["name", "email", "role"]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionAuth": []
    }
  ]
}